Add: Workspace Policy

+++ app/Http/Controllers/WorkspaceController.php
@@ -26,12 +26,14 @@ class WorkspaceController extends Controller
 
     public function show(Workspace $workspace)
     {
+        $this->authorize('view', $workspace);
         $task = $workspace->task;
         return view('workspace.show', compact('workspace','task'));
     }
 
     public function delete(Workspace $workspace)
     {
+        $this->authorize('delete', $workspace);
         $workspace->task()->delete();
         $workspace->delete();
 
@@ -40,11 +42,13 @@ class WorkspaceController extends Controller
 
     public function edit(Workspace $workspace)
     {
+        $this->authorize('update', $workspace);
         return view('workspace.update', compact('workspace'));
     }
 
     public function update(WorkspaceUpdateRequest $request, Workspace $workspace)
     {
+        $this->authorize('update', $workspace);
         $workspace->update([
             'user_id'=> auth()->user()->id,
             'name'=> $request-> name,
